+++ public/data/DetailPage/DetailPageHeader/DetailPage_Accommodations_Data.json
