+++ public/data/accommodations.json
