+++ src/pages/DetailPage/DetailPageHeader/DetailPageHeader.js
