+++ src/pages/DetailPage/DetailPageReservation/DetailPageReservationCheckInOut.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const DetailPageReservationCheckInOut = () => {
+  return (
+    <div>
+      <div>
+        <label for="checkInOut">체크인 / 체크아웃</label>
+        <input id="checkInOut" />
+      </div>
+    </div>
+  );
+};
+
+export default DetailPageReservationCheckInOut;
